Extract portfolio liquidity contracts
